+++ src/layout/index.js
@@ -5,7 +5,7 @@ import Modal from './Modal'
 import Popover from './Popover'
 import SpacedList from './SpacedList'
 import TextHighlight from './TextHighlight'
-import { ModalPicker } from './Pickers'
+import ModalPicker from './ModalPicker'
 import NestedPicker from './NestedPicker'
 import Dynamic from './Dynamic'
 import BarChart from './BarChart'
